Add class modules for 'Book' and 'Methods'
