+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./style.css" />
+    <link rel="stylesheet" href="./index.css" />
     <title>Awesome Books App</title>
   </head>
   <body>
@@ -49,6 +49,6 @@
       <p>Copyright ...</p>
     </footer>
 
-    <script src="./script.js"></script>
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
