+++ index.js
@@ -1,4 +1,7 @@
 /* eslint-disable max-classes-per-file */
+import Book from './modules/Book.js';
+import Methods from './modules/Methods.js';
+
 window.onload = () => {
   const add = document.querySelector('.add');
   const dateEl = document.querySelector('.date');
@@ -10,50 +13,6 @@ window.onload = () => {
     }, 1000);
   };
   displayDate();
-  class Book {
-    constructor(title, author) {
-      this.title = title;
-      this.author = author;
-    }
-  }
-  class Methods {
-    constructor() {
-      this.bookList = [];
-      this.addBook = document.querySelector('.all-books');
-      this.parser = new DOMParser();
-    }
-
-    removeBook = (e, newBookElement) => {
-      const index = e.target.getAttribute('myIndex');
-      this.bookList = this.bookList.filter((element, i) => {
-        if (i === parseInt(index, 10)) {
-          return false;
-        }
-        return true;
-      });
-      newBookElement.remove();
-      localStorage.setItem('bookArray', JSON.stringify(this.bookList));
-      this.showBooks();
-    };
-
-    showBooks = () => {
-      this.addBook.innerHTML = '';
-      this.bookList.forEach((e, i) => {
-        const newBook = `
-        <div class="book-div">
-        <p>"${e.title}" by ${e.author}</p>
-        <button type="button" class="remove" myIndex ="${i}" >Remove</button>
-        </div>
-        `;
-        const newBookElement = this.parser.parseFromString(newBook, 'text/html').body.firstChild;
-        const remove = newBookElement.querySelector('.remove');
-        remove.addEventListener('click', (e) => {
-          this.removeBook(e, newBookElement);
-        });
-        this.addBook.append(newBookElement);
-      });
-    };
-  }
 
   const method = new Methods();
   const bookStorage = localStorage.getItem('bookArray');
